casepilot-main/requirements.txt
google-adk>=1.0.0
google-cloud-firestore>=2.16.0
google-genai>=1.0.0
mcp>=1.0.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
